requirements.txt
flask==2.3.4
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.0
scikit-learn==1.3.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.2.0
langchain==0.1.103
